removed annoying OpERaTioNs File
